requirements.txt

numpy>=1.24
Pillow>=10.0
torch>=2.2; platform_system != "Windows" or platform_machine != "arm64"
